Adds bindings for converting C++ object pointers to and from foreign values.
(Also fixes HLint complaints; accidentally did this in the same commit.)

This provides explicit access to the conversion machinery performed when
passing values of type TObj a, and enables converting C++ object pointers
to Haskell values, and Haskell values to new C++ objects on the heap.

On the C++ side, we add genenc__{ClassName} and gendec__{ClassName}
functions for the C++-side encoding.  For Haskell, we add instances of
new typeclasses Encodable and Decodable.

The terminology is somewhat ambiguous.  In terms of an individual
language's bindings (C++, Haskell, whatever), "encoding" always refers to
converting a native value to its C type, and "decoding" vice versa.  But
the encode (resp. decode) function for a class always converts from a
foreign value to a C++ object (resp. C++ object to a foreign value).