Multi-modules: Initial working state.
Makes Cppop capable of generating multiple modules that may depend on each
other.  The specification hierarchy is now "interface has modules, modules
have exports."  Adds C++ requirements in the form of the 'Reqs' type on
Enum, Function, and Class, and further requirements for encoding within
CppCoder.  For Haskell, adds a (string) import set that encompases using
the Haskell-side and C-side types and encoding functions.  This generates
a ton of redundant imports that will be fixed in a future commit.  The
common Haskell generator code now takes care of generating module headers,
exports, and imports.