Add support for generating circular modules.
Cppop modules can depend on each other circularly.  This is something
that C++ allows e.g. with forward declarations to enable classes in
separate modules to have pointers to each other.

The Haskell code generator had problems with this, because GHC doesn't
allow circular modules without additional work.  Now the generator
detects strongly connected components, and for each module in a
nontrivial SCC, we also generate a hs-boot file akin to a C++ header
(that only contains type declarations, enums, things that belong in
header files), and switch all intra-component imports to be {-#SOURCE#-}
imports.

To support this, we have to include class hierarchies (via Haskell
typeclasses and instances) in the hs-boot file -- bindings need to be
able to upcast from source imports alone -- and because of this we have
to move methods out of typeclasses, otherwise we couldn't actually break
a lot of cycles.