Wrap std::vector; can't depend on type arguments in template conversions yet.
This creates a std::vector wrapper, and moves the class template conversion
code from StdList to StdVector, since it's doing random access to read the
elements to convert to a Haskell list.

This still doesn't quite work, because there's no way to depend on a
general type within the conversion generator, so that e.g. you can say:

  do [t] <- askTypeArgs  -- t :: Type
     name <- importType t
     ... "encodeAs (undefined :: " ++ name ++ ")" ...

The conversion code for all of these classes will be written properly once
Cppop can express the conversions needed.

This also adds "Encodable a a" instances for primitive types "a", and adds
accessor methods for chars in strings.