Prepend class names to method and ctor external names automatically.
Previously, method and ctor external names were completely separate from
class external names, and the two didn't have to be related in any way.
In practice, this was a pain because mkMethod, etc. then have to go
through the trouble of adding the class name to the method name.  It was
also causing problems with class templates, because the name of the
instantiated class is not known when the template is defined.

This switches to unconditionally prepending "className_" to all method
and ctor external names.  The class arguments from all mkMethod, mkCtor,
mkProp functions are gone now.