Add TObjToHeap for non-convertible C++ objects.
This fixes the problem of C++ code that passes non-convertible objects to
foreign code directly -- for example, functions returning iterators.
(TObjToHeap Class) is a new 'Type' data constructor that can only be used
in data leaving C++.  It takes an object of the given class, creates a new
copy on the heap, and passes a pointer to the copy to the foreign code.

Also adds withScopedPtr for defining the lifetime of a pointer in Haskell.