Fix broken casting for classes with multiple inheritance.
Up-casting was assuming that no pointer arithmetic needs to be done, which
is likely to break with classes using multiple inheritance.  Now, we
generate a C++ upcast function for every direct or indirect superclass of a
class, and use these to coerce arguments to the correct type prior to a C++
call.  This is quadratic in the number of generated bindings (with deep
hierarchies) and has constant processing time for a function call with
respect to the types of the arguments, rather than generating bindings only
for the direct superclasses of a class (linear) and doing
(aFromB . bFromC . cFromD) cast chains in function calls (linear).
Assuming that hierarchies are shallow and function call speed is more
important.

Const casting is still done entirely in Haskell, keeping the pointer the
same.

This commit also adds export lists to generated Haskell modules, because
generated Haskell binding code uses primitive cast functions from other
modules, and these functions are foreign imports, which are not exported
when a module has no export list.

Also, Haskell enum argument processing was using (/=) without importing it.