Add foreign garbage collection of C++ objects (issue #5).
This commit adds support for managing C++ objects with a foreign language's
garbage collector.  The generated Haskell data types that refer to C++ objects
are augmented to support being in a "managed" state (via a separate data
constructor) whereby the garbage collector is in charge of the object's
lifetime.  A new data constructor is added to Type:

    data Type = ... | TToGc Type

This constructor can be used only with object types being passed out of C++, and
causes ownership transfer.  (In the case of (TToGc . TObj), the object is copied
to the heap first, it is not decoded as with TObj.)  See the documentation for
TToGc and the new methods in CppPtr.

Details about the required changes:

In Haskell, changing the representation of pointers from newtypes to data types
with multiple constructors means that they can't be used directly in foreign
import statements any more.  I opted to make the C-side types for object
pointers have Ptr in front.  This caused weirdness because the Foo data type
"is" a pointer to a Foo, but a Ptr Foo also "is" a pointer to a Foo (and not a
double pointer!), so I removed the (dervied) Storable Foo instance.  To use
peek/poke you now have to convert to pointers manually with 'toPtr'.

I also made the Eq and Ord instances for object pointer types only consider the
pointer address, not the managed state, so a managed pointer will be equal to an
unmanaged one when they point to the same object.