Enable passing TObj of non-convertible object types out of C++.
This adds options to a class's language-specific conversion settings, so
that when Hoppy sees (TObj cls) being returned from a C++ function or being
passed into a foreign callback, Hoppy can automatically interpret that to
mean TObjToHeap or TToGc . TObj.  This is particularly useful for using
mkProp with such types.