Initial C++ exception support (issue #10).
Exceptions can be thrown from C++ and caught by Haskell.  Catching is supported
for classes explicitly declared as being used for exceptions.  Additionally,
(...) may be caught.  All C++ functions that throw exceptions must have the
exceptions they throw declared, so that Hoppy can generate the proper glue code.

Catching is done with a separate catchCpp function, which implements logic for
catching supertypes of exception classes.

Throwing from Haskell is not supported yet.

This also adds an Ord instance for Class, and makes this instance as well as the
existing Eq instances for Class, CppEnum, and Variable use external names
instead of identifiers (since these are guaranteed to be unique, whereas
identifiers aren't, and there's no Ord instance for Identifier).