Unify class entities (ctors, methods, properties).
Previously, constructors and methods went into separate fields in a Class,
and this distiction carried over to the declaration of user classes, and
required special list munging to handle properties.

Now, a Class simply has a list of ClassEntity, which can be Ctor, Method,
or Prop.  These entities can be specified in any order, and properties can
be listed directly with mkProp ("++ mkProps" isn't needed anymore).