Make function pointers work: ptrT (fnT ...).
This takes the part of the callback machinery that creates FunPtrs from Haskell
functions and exposes it.  Function pointers always use C-side types for their
parameters and arguments, whereas previously C-side types were not public.  This
is done because wanting to support arbitrary function pointers at runtime
conflicts with having a C++ function pointer that must call a function at a
fixed address.