Add exception throwing from Haskell to C++ (closes #10).
C++ exceptions can now be thrown from Haskell, and either recaught in Haskell,
or passed on to C++ when a callback is marked as throwing.  This turned out
better than C++-to-Haskell transfer: for callbacks, we don't have to list out
each potential exception class thrown.  The exception support module now needs
to be specified manually, since there is C++ support code now too, and
preventing circular dependencies amongst C++ modules is a manual process.

Exceptions are documented at the end of the user's guide.

This also adds toHsCtorName to match toHsMethodName.