Massive refactor (issues #4, #7, #9, #12, #17, #30).
TODO Lots of docs and todos.
TODO Emit documentation showing function parameter names (also show objToHeapT and toGcT).

This addresses a number of long-standing issues with Hoppy:
- #4  [feat] Support arbitrary enum values at runtime
- #7  [bug]  Incorrect Haskell instances for non-default-valued enums
- #9  [feat] Generalize numeric type definitions
- #12 [feat] Optional function parameter names
- #17 [feat] Autodetection for enum values
- #30 [feat] Enums should support types other than 'int'

The major changes in this refactor are:

1. Generalize data types (issue #9).

Previously, variables, enums (bitspaces), functions, and classes were all
special things that were built-in to Hoppy itself quite deeply.  They had
their own Type and Export data constructors, with their logic split between
lots of 'case' statements throughout the generator.

All of this special handling has been removed.  In its place, there is now
a generic Exportable typeclass of which Function, Class, etc. are
instances.  The C++ and Haskell generators each call a method on this
typeclass in order to produce code.  Each of these entity types has been
split into its own module.

Replacing the myriad of Type data constructors is a single TManual
constructor, that takes a new ConversionSpec data type defining everything
necessary to perform conversions for a specific type of value.  enumT, etc.
now return TManual.  Exceptions to this are fnT, which is not convertible
directly, and objT, which is still handled specially (this will be
refactored in the future).

The separate Type data constructors for C++ numeric types are no longer
built-in.  Custom ones can be created with 'makeNumericType'.

2. Autodetection of enum values (issue #17).

One big pain point with writing Hoppy bindings is the need to manually
specify numeric values for enum entries.  It's tolerable for libraries like
Qt that document these values, but it's painful if you have to figure them
out yourself.

Now, using 'makeAutoEnum' instead of 'makeEnum', a generator will build a
C++ program when run in order to determine what the value for each entry
is, directly from the compiler.  This does require a compiler to be
available.  The default is G++, but others can be used.

This function also comes with easier input for enum entry names.  C++ entry
names can be typed directly, and they are split into words for Haskell
automatically, rather than having to ["split", "them", "yourself"].

3. Address enum issues (issues #4, #7, #30).

This fixes a few problems and limitations with the implementation of enums.

First, only enum values that were defined in the interface were supported.
If new ones were added, or even if the binding was simply incomplete to
start with, gateway code would crash when it received a value it didn't
know about.  Now, enums may have a "unknown" Haskell data constructor that
will be used as a fallback.  Enums are given one by default, but this can
be disabled on a per-enum basis.

Second, the enums in Haskell received Bounded and Enum instances, that
didn't work well in the face of unknown entries, and the default Enum
methods didn't work when the numeric values weren't all in sequence.  Enums
no longer get these two instances.  Instead, they get a new Hoppy-specific
CppEnum instance.  This can be expanded as necessary.

Third, Hoppy assumed that enums were backed by a C++ int.  This is bad, it
can be wrong on both size and signedness.  Now, manually-specified enums
can set their types directly, but normally this will be detected
automatically while detecting enum values (and additionally
explicitly-provided types are validated by default).

Also, enums in Haskell now get Bits instances.  This is on by default and
can be disabled.

4. Remove bitspaces.

Bitspaces are gone.  They were a confusing split between enums that were
hacked into Hoppy in order to support the QFlags wrapper that Qt uses in
places where multiple values in an enum are accepted.  This has been
reimplemented in Qtah directly as "flags", and the functionality that
bitspaces added on top of enums (custom numeric types; Bits instance) has
moved to enums.

5. Support function parameter names (issue #12).

Names and documentation strings can now be attached to function parameters.
Currently this isn't used anywhere, but it will be used to generate
documentation for generated bindings.

**This does cause API breakage for packages using Hoppy.**

Parameter lists to 'makeFn', 'mkMethod', etc. are now generic.  They can
either be [Type] as before, or (~:) can be used to attach names to types to
give [Parameter].  The breakage is that the empty list is now ambiguously
typed, so a new 'np = [] :: [Parameter]' value should be used instead.

6. Other

There are new --dump-ext-names and --dump-enums options to the generator command
line interface.  These list out assembled information about the interface for
debugging purposes.  For enums, individual entries are listed, with numeric
values determined by C++ compiler, if applicable.