add regHook for *-cpp package to copy shared library file
A package can be registered 'in-place' without performing copy action
or install action. We need to make sure that the shared library file
is copied in such case.

I encountered such case when I tried to use 'cabal new-build'.
I'm not sure if this is correct way to do, but it works fine in my
environment.