Split set iterator types into const and non-const versions.
I'd made a bad assumption when writing the std::set binding originally;
it looked like set::iterator and set::const_iterator were the same type,
and while they may be in some cases, they appear to be different for
unordered_set, with g++ 7.2.0 on Kubuntu 17.10 amd64.

So this creates separate bindings for iterator and const_iterator for each
of these set types, and separate const and non-const begin() and end()
methods.