Add documentation; improve ConversionSpecCpp logic with default behaviour.
This adds lots of documentation, particularly around ConversionSpecs.  This also
revises the behaviour for ConversionSpecCpps to have default conversion
behaviour in the absense of conversionSpecCpp{From,To}CppExpr, when no explicit
conversion logic is necessary.  This also moves fnT and fnT' from Types.hs to
Function.hs to match how the other types with modules have their Type
constructors there.