Compatibility with GHC 8.6, Cabal 2.4.
This includes a fix for a failable pattern.  As part of the MonadFail
Proposal, this now requires MonadFail.  The code was reworked not to use
list destructuring.

Unit tests that counted PtrCtr constructions and destructions were broken
with respect to garbage collection, because PtrCtrs created in previous
tests could contribute to destruction counts in the current test.  Don't
know why this didn't manifest before, but it does in a big way with GHC 8.6
(four tests failed).  The tests were fixed to always wait for leftover
instances to be cleaned up.