Use buildStrictly; clean up many warnings; add -full.nix expressions.
This enables buildStrictly for Nix builds.  This is a Nixpkgs function that
builds packages from sdists rather than a copy of the current source
directory (great!), and also enables -Wall -Werror.

I cleaned up a bunch of the warnings as suggested by -Wall, mainly name
shadowing and type signatures for top-level bindings, but at the end of the
day disabled the -Wall part of buildStrictly, because some warnings are
annoying (-Wtype-defaults).  We implement buildStrictly ourselves in
build-strictly.nix.

This also splits the Nix package sets into regular and full versions.  The
regular versions build against the current GHC in Nixpkgs, and the full
versions build against all of the verisons defined in haskells.nix as
before.

Thanks to building from sdist, this adds a few missing files to
tests/suites/*/cpp/*.cabal